Add unit tests for pedigree loading and role guessing
Added unit test to assert that guess_role_nuc does not modify the input
dataframe
Moved unit tests from test_pedigree_reader to test_pedigree_loading
Added unit test for loading a pedigree file without a header (using
column indices)
