+++ dae/dae/pedigrees/tests/test_role_guessing.py
@@ -8,10 +8,15 @@ from dae.variants.attributes import Role
 
 
 def test_role_guessing(sample_nuc_family):
+    ped_df_copy = PedigreeReader.load_pedigree_file(sample_nuc_family)
     ped_df = PedigreeReader.load_pedigree_file(sample_nuc_family)
     assert ped_df is not None
 
-    ped_df = PedigreeRoleGuesser.guess_role_nuc(ped_df)
-    assert ped_df is not None
-    assert 'role' in ped_df
-    assert list(ped_df['role']) == [Role.dad, Role.mom, Role.prb, Role.sib]
+    res_df = PedigreeRoleGuesser.guess_role_nuc(ped_df)
+    assert res_df is not None
+    assert 'role' in res_df
+    assert list(res_df['role']) == [Role.dad, Role.mom, Role.prb, Role.sib]
+
+    # Assert lack of side effects
+    assert 'role' not in ped_df  # Check for side-effects
+    assert ped_df.equals(ped_df_copy)
